+++ pyannote/audio/pipelines/speech_separation.py
@@ -654,6 +654,12 @@ class SpeechSeparation(SpeakerDiarizationMixin, Pipeline):
                 sources.data * discrete_diarization.align(sources).data[:, :num_sources]
             )
 
+        # separated sources might be scaled up/down due to SI-SDR loss used when training
+        # so we peak-normalize them
+        sources.data = sources.data / np.max(
+            np.abs(sources.data), axis=0, keepdims=True
+        )
+
         # convert to continuous diarization
         diarization = self.to_annotation(
             discrete_diarization,